refactor(cloud-ui): collapse 6 use-session-auth + 3 auth-query + 2 auth-gate clones into cloud/lib (#11341)
~1,000 lines of cloned Steward session/auth-gate glue existed across
billing, api-explorer, instances, public-pages, account-security,
monetization, api-keys, and analytics. No production consumer used the
divergent extended fields (authSource/stewardAuthenticated/stewardUser)
— only a test mock did. One canonical hook now lives in cloud/lib:
use-session-auth (with the Playwright test-auth bypass the browser
suites rely on) + auth-query.

Refs #11341
