+++ packages/ui/src/cloud/instances/AgentDetailPage.tsx
@@ -33,7 +33,7 @@ import { ElizaConnectButton } from "./components/eliza-connect-button";
 import { useAgent } from "./lib/data/eliza-agents";
 import { useT } from "./lib/i18n";
 import { statusBadgeColor, statusDotColor } from "./lib/sandbox-status";
-import { useRequireAuth } from "./lib/use-session-auth";
+import { useRequireAuth } from "../lib/use-session-auth";
 
 function formatDate(date: string | null): string {
   if (!date) return "—";
