+++ packages/ui/src/cloud/public-pages/pages/payment/payment-success-page.tsx
@@ -10,7 +10,7 @@ import { CheckCircle, Loader2 } from "lucide-react";
 import { useEffect } from "react";
 import { useNavigate, useSearchParams } from "react-router-dom";
 import { useCloudT } from "../../../shell/CloudI18nProvider";
-import { useSessionAuth } from "../../lib/use-session-auth";
+import { useSessionAuth } from "../../../lib/use-session-auth";
 
 export default function PaymentSuccessPage() {
   const t = useCloudT();
